Fix player "bouncing" up and down ~0.5 pixels per frame
Previously there was a bug in the fall detection code (simplified
below):

    if onGround and yVelocity > 0 then
        yVelocity = 0
        yAccel = 0
    else
        yAccel = fallAccel
    end

This meant that if the player was on the ground and *not* currently
falling, it should start falling. The next frame, the player would be on
the ground and falling, and the fall would stop.

The code that sets the player back on top if it is underground caused a
jump back up. Without that code, the player would slowly sink into the
ground as if sinking into sand.

I solved this by adding another state into the animation state machine.
Previously the state machine looked like this:

+------+
|      +<------------------
|stand |                  |
+---+--+                  |
    |                     |
    |                     |
    |                     |
    |                     |
if on ground          if on ground
    |                     |
    |                     |
    V                     |
+---+--+                  |
|      +-------------------
| jump |
+------+

Now it looks like this:

+------+
|      +<------------------
|stand |                  |
+---+--+                  |
    |                     |
    |                     |
    |                     |
    |                     |
if on ground          if on ground
    |                     |
    |                     |
    V                     |
+---+--+     if       +---+--+
|      +--yVelocity-->+      |
| jump |     >0       | fall |
+------+              +------+