Fix player falling underground for 1 frame
Previously, the player's feet would be underground for 1 frame after
falling and landing on the ground. This happened because the change in
position would be applied, then the player would be drawn, then the
player would be placed on top of the ground at the beginning of the next
update iteration. Now the replacement happens immediately after setting
the position.